переносит сабмит в модуль формы
