+++ js/api-fetch.js
@@ -1,4 +1,3 @@
-import {isEscapeKey} from './utils.js';
 const INPUT_DATA_SOURCE = 'https://27.javascript.pages.academy/kekstagram-simple/data';
 const OUTPUT_DATA_SOURCE = 'https://27.javascript.pages.academy/kekstagram-simple';
 
@@ -20,43 +19,21 @@ const getImageData = (onSuccess, onError) => {
 
 export {getImageData};
 
-// const onModalEcsKeydown = (evt, child) => {
-//   if (isEscapeKey(evt)) {
-//     evt.preventDefault();
-//     document.body.removeChild(child);
-//   }
-// };
-
 const sendImageData = (onSuccess, onError, body) => {
-  fetch(
+  fetch (
     OUTPUT_DATA_SOURCE,
     {
       method: 'POST',
-      body
-    },
-  )
+      body,
+    })
     .then((response) => {
       if (response.ok) {
-        closeModal();
-        const successFragment = document.createDocumentFragment();
-        const success = successTemplate.cloneNode(true);
-        successFragment.appendChild(success);
-        document.body.appendChild(successFragment);
+        return onSuccess();
       }
-      else {
-        closeModal();
-        const errorFragment = document.createDocumentFragment();
-        const error = errorTemplate.cloneNode(true);
-        errorFragment.appendChild(error);
-        document.body.appendChild(errorFragment);
-      }
+      throw new Error();
     })
     .catch(() => {
-      closeModal();
-      const errorFragment = document.createDocumentFragment();
-      const error = errorTemplate.cloneNode(true);
-      errorFragment.appendChild(error);
-      document.body.appendChild(errorFragment);
+      onError();
     });
 };
